remove unused functions; fix some compilation warnings;

+++ Reference_Implementation/lib/bench/bench_cf.c
@@ -1,7 +1,5 @@
 #include <stdio.h>
 #include <math.h>
-#include <stdlib.h>
-#include <string.h>
 
 #include "canonical.h"
 #include "cycles.h"
@@ -14,8 +12,8 @@
 void normalized_rng(normalized_IS_t *V) {
     randombytes((uint8_t *)V->values, K*(N-K));
     const uint8_t mask = 0x7F;
-    for (uint32_t i = 0; i < K; ++i) {
-        for (uint32_t j = 0; j < K; ++j) {
+    for (unsigned i = 0; i < K; ++i) {
+        for (unsigned j = 0; j < K; ++j) {
             V->values[i][j] &= mask;
         }
     }
@@ -25,9 +23,9 @@ int bench_cf3(void) {
     normalized_IS_t G1, G2;
 
 	printf("cf3:\n");
-    uint64_t c = 0, c1, ctr = 0;
+    unsigned c = 0, c1, ctr = 0;
 	normalized_sf(&G2);
-    for (uint64_t i = 0; i < ITERS; i++) {
+    for (unsigned i = 0; i < ITERS; i++) {
 		normalized_copy(&G1, &G2);
 
     	c -= read_cycle_counter();
@@ -35,10 +33,10 @@ int bench_cf3(void) {
         c += read_cycle_counter();
     }
     c1 = c/ITERS;
-    printf("non-ct: %ld cyc, ctr: %ld\n", c1, ctr);
+    printf("non-ct: %u cyc, ctr: %u\n", c1, ctr);
 
     c = 0; ctr = 0;
-    for (uint64_t i = 0; i < ITERS; i++) {
+    for (unsigned i = 0; i < ITERS; i++) {
         normalized_copy(&G1, &G2);
 
         c -= read_cycle_counter();
@@ -47,7 +45,7 @@ int bench_cf3(void) {
     }
 
     c = c/ITERS;
-    printf("ct: %ld cyc, ctr: %ld\n", c, ctr);
+    printf("ct: %u cyc, ctr: %u\n", c, ctr);
     printf("factor %lf\n\n", (double)c/(double)c1);
     return 0;
 }
@@ -57,8 +55,8 @@ int bench_cf4(void) {
 
 	printf("cf4:\n");
 	normalized_sf(&G2);
-    uint64_t c = 0, c1, ctr = 0;
-    for (uint64_t i = 0; i < ITERS; i++) {
+    unsigned c = 0, c1, ctr = 0;
+    for (unsigned i = 0; i < ITERS; i++) {
 		normalized_copy(&G1, &G2);
 
         c -= read_cycle_counter();
@@ -66,10 +64,10 @@ int bench_cf4(void) {
         c += read_cycle_counter();
     }
     c1 = c/ITERS;
-    printf("non ct: %ld cyc, ctr: %ld\n", c1, ctr);
+    printf("non ct: %u cyc, ctr: %u\n", c1, ctr);
 
     c = 0; ctr = 0;
-    for (uint64_t i = 0; i < ITERS; i++) {
+    for (unsigned i = 0; i < ITERS; i++) {
 		normalized_copy(&G1, &G2);
 
         c -= read_cycle_counter();
@@ -78,7 +76,7 @@ int bench_cf4(void) {
     }
 
     c = c/ITERS;
-    printf("ct: %ld cyc, ctr: %ld\n", c, ctr);
+    printf("ct: %u cyc, ctr: %u\n", c, ctr);
     printf("factor %lf\n\n", (double)c/(double)c1);
     return 0;
 }
@@ -88,8 +86,8 @@ int bench_cf5(void) {
 
 	normalized_sf(&G2);
 	printf("cf5:\n");
-    uint64_t c = 0, c1, ctr = 0;
-    for (uint64_t i = 0; i < ITERS; i++) {
+    unsigned c = 0, c1, ctr = 0;
+    for (unsigned i = 0; i < ITERS; i++) {
 		// normalized_copy(&G1, &G2);
         normalized_rng(&G1);
 
@@ -98,11 +96,11 @@ int bench_cf5(void) {
         c += read_cycle_counter();
     }
     c1 = c/ITERS;
-    printf("non ct: %ld cyc, ctr: %ld\n\n", c1, ctr);
+    printf("non ct: %u cyc, ctr: %u\n\n", c1, ctr);
 
 
     c = 0; ctr = 0;
-    for (uint64_t i = 0; i < ITERS; i++) {
+    for (unsigned i = 0; i < ITERS; i++) {
         // normalized_copy(&G1, &G2);
         normalized_rng(&G1);
 
@@ -112,12 +110,12 @@ int bench_cf5(void) {
     }
 
     c = c/ITERS;
-    printf("pop: %ld cyc, ctr: %ld\n", c, ctr);
+    printf("pop: %u cyc, ctr: %u\n", c, ctr);
     printf("factor %lf\n\n", (double)c/(double)c1);
 
 
     c = 0; ctr = 0;
-    for (uint64_t i = 0; i < ITERS; i++) {
+    for (unsigned i = 0; i < ITERS; i++) {
         // normalized_copy(&G1, &G2);
         normalized_rng(&G1);
 
@@ -127,7 +125,7 @@ int bench_cf5(void) {
     }
 
     c = c/ITERS;
-    printf("fast: %ld cyc, ctr: %ld\n", c, ctr);
+    printf("fast: %u cyc, ctr: %u\n", c, ctr);
     printf("factor %lf\n\n", (double)c/(double)c1);
     return 0;
 }
